pyproject.toml
[project]
name = "autodb-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=4.0.0",
    "boto3>=1.40.21",
    "fastapi>=0.116.1",
    "motor>=3.7.1",
    "pymongo>=4.14.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "uvicorn[standard]>=0.35.0",
]

